Extract the adapter to its own file
